third_party/blink/web_tests/platform/generic/external/wpt/fetch/private-network-access/service-worker-update.https.window-expected.txt

This is a testharness.js-based test.
FAIL update public to local: failed preflight. assert_equals: update error expected (string) "TypeError" but got (undefined) undefined
FAIL update public to local: success. assert_equals: update error expected (undefined) undefined but got (string) "TypeError"
FAIL update public to private: failed preflight. assert_equals: update error expected (string) "TypeError" but got (undefined) undefined
FAIL update public to private: success. assert_equals: update error expected (undefined) undefined but got (string) "TypeError"
PASS update public to public: success.
Harness: the test ran to completion.
